+++ src/core/mod.rs
@@ -1,4 +1,7 @@
-pub mod byte;
-pub mod double;
+mod byte;
+mod double;
 pub mod memory;
 pub mod consts;
+
+pub use byte::Byte;
+pub use double::Double;
