+++ src/mapper/mapper_nrom.rs
@@ -1,0 +1,5 @@
+use super::Mapper;
+
+pub struct NROMMapper {
+
+}
