--- src/rom_parser/errors.rs
@@ -1,4 +1,0 @@
-#[derive(Debug)]
-pub enum ParserError {
-    InvalidRom
-}
