+++ src/rom_parser/mod.rs
@@ -1,2 +1,6 @@
 pub mod ines;
-pub mod errors;
+
+#[derive(Debug)]
+pub enum ParserError {
+    InvalidRom
+}
